requirements.txt
python-telegram-bot==20.7
Flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0